pyproject.toml
[tool.poetry]
name = "cmyr-cdk"
version = "0.1.0"
description = ""
authors = ["c7c8 <tothemyers@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"aws-cdk.core" = "^1.97.0"
"aws-cdk.aws-s3" = "^1.97.0"
"aws-cdk.aws-route53" = "^1.97.0"
"aws-cdk.aws-cloudfront" = "^1.97.0"
"aws-cdk.aws-cloudfront-origins" = "^1.97.0"
"aws-cdk.aws-certificatemanager" = "^1.97.0"
"aws-cdk.aws-route53-targets" = "^1.97.0"
"aws-cdk.aws-codebuild" = "^1.97.0"
"aws-cdk.aws-iam" = "^1.97.0"
"aws-cdk.aws-sns" = "^1.97.0"
"aws-cdk.aws-sns-subscriptions" = "^1.97.0"
"aws-cdk.aws-events-targets" = "^1.97.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
